requirements.txt
aiohttp
anytree
appdirs
https://github.com/Anshusharma75/jhamak/archive/refs/heads/master.zip
attrdict
beautifulsoup4
cfscrape
feedparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
js2py
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
pyshorteners
python-dotenv
python-magic
python-telegram-bot
https://github.com/Anshusharma75/gandu/archive/refs/heads/master.zip
requests
https://github.com/Anshusharma75/jati/archive/refs/heads/master.zip
telegraph
tenacity
TgCrypto
https://github.com/hsjsa/tttool/archive/refs/heads/master.zip
urllib3
https://github.com/Anshusharma75/prabhu/archive/refs/heads/master.zip
heroku3